Separating list logic into separate component from popover container
